# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: WB Analytics API
  version: 1.0.0
  description: Контракт API сервиса аналитики товаров Wildberries. Снят с Django-реализации
    и заморожен на время миграции.
paths:
  /api/products/:
    get:
      operationId: products_list
      summary: List products
      parameters:
      - in: query
        name: max_price
        schema:
          type: number
          format: double
          minimum: 0
        description: Upper bound on the discounted price.
      - in: query
        name: max_rating
        schema:
          type: number
          format: double
          minimum: 0
          maximum: 5
        description: Upper bound on rating.
      - in: query
        name: max_reviews
        schema:
          type: integer
          minimum: 0
        description: Upper bound on the review count.
      - in: query
        name: min_price
        schema:
          type: number
          format: double
          minimum: 0
        description: Lower bound on the discounted price.
      - in: query
        name: min_rating
        schema:
          type: number
          format: double
          minimum: 0
          maximum: 5
        description: Lower bound on rating.
      - in: query
        name: min_reviews
        schema:
          type: integer
          minimum: 0
        description: Lower bound on the review count.
      - in: query
        name: ordering
        schema:
          type: string
          enum:
          - -name
          - -price
          - -rating
          - -reviews_count
          - -sale_price
          - name
          - price
          - rating
          - reviews_count
          - sale_price
        description: 'Sort field, optionally prefixed with ''-'' for descending. One
          of: name, price, rating, reviews_count, sale_price.'
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
        description: 1-based page number.
      - in: query
        name: page_size
        schema:
          type: integer
          minimum: 1
        description: Rows per page. Values above 1000 are clamped to 1000 rather than
          refused; the cap exists because the charts are computed client-side from
          this response.
      - in: query
        name: query
        schema:
          type: string
        description: Restrict to products collected for this search query.
      tags:
      - products
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/parse/:
    post:
      operationId: collections_create
      description: POST /api/parse/ — enqueue an async collection run, returns 202.
      summary: Enqueue a collection run
      tags:
      - parse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseRequestRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseEnqueued'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/tasks/{task_id}/:
    get:
      operationId: collections_retrieve
      description: GET /api/tasks/{task_id}/ — async collection status.
      summary: Collection run status
      parameters:
      - in: path
        name: task_id
        schema:
          type: string
        required: true
      tags:
      - tasks
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseJob'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/products/{wb_id}/history/:
    get:
      operationId: products_price_history
      description: GET /api/products/{wb_id}/history/ — a product's price time-series.
      summary: Price history of one product
      parameters:
      - in: path
        name: wb_id
        schema:
          type: integer
        required: true
      tags:
      - products
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceHistory'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/stats/:
    get:
      operationId: stats_retrieve
      description: 'One `query` returns a Stats object. Two or more return `{items:
        [{query, stats}]}` — the same path answers with a different shape, which is
        why both are declared.'
      summary: Aggregates for a filtered set, or a comparison of queries
      parameters:
      - in: query
        name: max_price
        schema:
          type: number
          format: double
          minimum: 0
        description: Upper bound on the discounted price.
      - in: query
        name: max_rating
        schema:
          type: number
          format: double
          minimum: 0
          maximum: 5
        description: Upper bound on rating.
      - in: query
        name: max_reviews
        schema:
          type: integer
          minimum: 0
        description: Upper bound on the review count.
      - in: query
        name: min_price
        schema:
          type: number
          format: double
          minimum: 0
        description: Lower bound on the discounted price.
      - in: query
        name: min_rating
        schema:
          type: number
          format: double
          minimum: 0
          maximum: 5
        description: Lower bound on rating.
      - in: query
        name: min_reviews
        schema:
          type: integer
          minimum: 0
        description: Lower bound on the review count.
      - in: query
        name: query
        schema:
          type: array
          items:
            type: string
        description: Repeat to compare queries side by side.
      tags:
      - stats
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsOrComparison'
          description: 'Single query: Stats. Repeated query: {items: [{query, stats}]}.'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/export/:
    get:
      operationId: export_download
      description: |-
        GET /api/export/?format=csv|xlsx — the filtered product set as a file.

        Authenticated-only: this dumps the whole filtered catalogue in one request, so
        it is an internal tool rather than part of the public catalogue read surface.
        Throttled for the same reason.

        ``format`` stays a plain query parameter because DRF's renderer-suffix lookup
        is disabled project-wide (``URL_FORMAT_OVERRIDE = None``).
      summary: Download the filtered set as a file
      parameters:
      - in: query
        name: format
        schema:
          type: string
          enum:
          - csv
          - xlsx
        description: Defaults to csv.
      - in: query
        name: max_price
        schema:
          type: number
          format: double
          minimum: 0
        description: Upper bound on the discounted price.
      - in: query
        name: max_rating
        schema:
          type: number
          format: double
          minimum: 0
          maximum: 5
        description: Upper bound on rating.
      - in: query
        name: max_reviews
        schema:
          type: integer
          minimum: 0
        description: Upper bound on the review count.
      - in: query
        name: min_price
        schema:
          type: number
          format: double
          minimum: 0
        description: Lower bound on the discounted price.
      - in: query
        name: min_rating
        schema:
          type: number
          format: double
          minimum: 0
          maximum: 5
        description: Lower bound on rating.
      - in: query
        name: min_reviews
        schema:
          type: integer
          minimum: 0
        description: Lower bound on the review count.
      - in: query
        name: ordering
        schema:
          type: string
          enum:
          - -name
          - -price
          - -rating
          - -reviews_count
          - -sale_price
          - name
          - price
          - rating
          - reviews_count
          - sale_price
        description: 'Sort field, optionally prefixed with ''-'' for descending. One
          of: name, price, rating, reviews_count, sale_price.'
      - in: query
        name: query
        schema:
          type: string
        description: Restrict to products collected for this search query.
      tags:
      - export
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
...
components:
  schemas:
...
    Error:
      type: object
      description: |-
        The `{"detail": ...}` body the shared exception handler returns.

        Not the only error body in the API: DRF answers its own serializer
        validation with a field-keyed map instead, and the handler passes that
        through untouched. The docstring here used to claim this shape covered
        "every error path", which is what a contract test disproved — see
        `FIELD_ERRORS_SCHEMA` below and `VALIDATION_ERROR_RESPONSE`.
      properties:
        detail:
          type: string
      required:
      - detail
...
    ParseEnqueued:
      type: object
      description: '`POST /api/parse/` — 202 with the accepted job, not the finished
        one.'
      properties:
        task_id:
          type: string
        query:
          type: string
        status:
          type: string
      required:
      - query
      - status
      - task_id
    ParseJob:
      type: object
      properties:
        task_id:
          type: string
        query:
          type: string
        status:
          type: string
        created:
          type: integer
        updated:
          type: integer
        collected_count:
          type: integer
        error:
          type:
          - string
          - 'null'
        finished_at:
          type:
          - string
          - 'null'
          format: date-time
      required:
      - collected_count
      - created
      - error
      - finished_at
      - query
      - status
      - task_id
      - updated
    ParseRequestRequest:
      type: object
      description: '`POST /api/parse/` body. Validated by hand in the view, described
        here.'
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 200
        max_pages:
          type:
          - integer
          - 'null'
          maximum: 20
          minimum: 1
          description: Defaults to the server setting.
      required:
      - query
...
    PriceHistory:
      type: object
      description: '`GET /api/products/{wb_id}/history/` — the price time series.'
      properties:
        wb_id:
          type: integer
        points:
          type: array
          items:
            $ref: '#/components/schemas/Snapshot'
      required:
      - points
      - wb_id
    ProductPage:
      type: object
      description: '`GET /api/products/` — DRF''s page envelope, assembled manually.'
      properties:
        count:
          type: integer
        next:
          type:
          - string
          - 'null'
        previous:
          type:
          - string
          - 'null'
        results:
          type: array
          items:
            $ref: '#/components/schemas/ProductView'
      required:
      - count
      - next
      - previous
      - results
    ProductView:
      type: object
      properties:
        wb_id:
          type: integer
        name:
          type: string
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        sale_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        discount_abs:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        discount_pct:
          type: string
          format: decimal
          pattern: ^-?\d{0,4}(?:\.\d{0,2})?$
        rating:
          type: string
          format: decimal
          pattern: ^-?\d{0,1}(?:\.\d{0,1})?$
        reviews_count:
          type: integer
        query:
          type:
          - string
          - 'null'
        updated_at:
          type:
          - string
          - 'null'
          format: date-time
      required:
      - discount_abs
      - discount_pct
      - name
      - price
      - query
      - rating
      - reviews_count
      - sale_price
      - updated_at
      - wb_id
    QueryComparison:
      type: object
      description: |-
        `GET /api/stats/?query=a&query=b` — the comparison branch.

        Two shapes behind one path: a single `query` returns Stats directly, two or
        more return this. The schema has to say so, or a client generated from it
        breaks on the second query parameter.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QueryStats'
      required:
      - items
    QueryStats:
      type: object
      properties:
        query:
          type: string
        stats:
          $ref: '#/components/schemas/Stats'
      required:
      - query
      - stats
...
    Snapshot:
      type: object
      properties:
        captured_at:
          type: string
          format: date-time
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        sale_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        rating:
          type: string
          format: decimal
          pattern: ^-?\d{0,1}(?:\.\d{0,1})?$
      required:
      - captured_at
      - price
      - rating
      - sale_price
    Stats:
      type: object
      properties:
        count:
          type: integer
        avg_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        median_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        price_stddev:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        avg_discount_abs:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        discount_share:
          type: number
          format: double
        top_by_reviews:
          type: array
          items:
            $ref: '#/components/schemas/TopProduct'
      required:
      - avg_discount_abs
      - avg_price
      - count
      - discount_share
      - median_price
      - price_stddev
      - top_by_reviews
    StatsOrComparison:
      oneOf:
      - $ref: '#/components/schemas/Stats'
      - $ref: '#/components/schemas/QueryComparison'
...
    TopProduct:
      type: object
      properties:
        wb_id:
          type: integer
        name:
          type: string
        reviews_count:
          type: integer
      required:
      - name
      - reviews_count
      - wb_id
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT